# lib/api-zod/src/generated/types/lead.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Lead {
  id: number;
  email: string;
  phone?: string | null;
  businessType?: string | null;
  businessDescription?: string | null;
  website?: string | null;
  demoNeeds?: string | null;
  demoType?: string | null;
  consentGiven?: boolean | null;
  status?: string | null;
  createdAt?: string | null;
  updatedAt?: string | null;
}