# lib/api-zod/src/generated/types/updateLeadInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateLeadInput {
  phone?: string;
  businessType?: string;
  businessDescription?: string;
  website?: string;
  demoNeeds?: string;
  demoType?: string;
  consentGiven?: boolean;
  status?: string;
}